Code quality, refactoring, and update start.sh

+++ start.sh
@@ -1,3 +1,7 @@
 #!/usr/bin/env bash
-cp bin/ngrok.exe node_modules/ngrok/bin/ngrok.exe
+if [ ! -f node_modules/ngrok/bin/ngrok.exe ];
+then
+    cp bin/ngrok.exe node_modules/ngrok/bin/ngrok.exe
+    echo "Copied ngrok executable"
+fi
 node src/mari-bot.js -l
